requirements.txt
# API 客户端
anthropic>=0.25.0
openai>=1.30.0

# 视觉/视频处理
opencv-python>=4.9.0
Pillow>=10.0.0

# 音频处理
openai-whisper>=20231117
edge-tts>=6.1.9

# 工具
pyyaml>=6.0
tqdm>=4.66.0
